testing/web-platform/tests/workers/WorkerLocation_hash_nonexist.htm

<!DOCTYPE html>
<html>
<head>
<title> WorkerLocation.hash with no &lt;fragment&gt; component </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>
    
    var FileName = './support/WorkerLocation.js';
    var HashString = "";
    var ExpectedResult = HashString;

    var t = async_test("Test Description: WorkerLocation hash attribute returns an empty string when there is no &lt;query&gt; component in input URL.");
    
    var worker = new Worker(FileName + HashString);

    worker.onmessage = t.step_func(function(evt)
    {
        assert_equals(evt.data.hash, ExpectedResult);
        t.done();
    });
    
</script>
</body>
</html>
